Set a default value for the `spectral_radius`
While this is of questionable value in the real world, it makes it a tad
easier to create an exemplary `Esn` to play with.

Furthermore it visually separates what I consider acceptable positional
arguments (the structure defining `in_size`, `reservoir_size` and
`out_size`) from tuning parameters that should definitely be passed as
keyword args (although I will admit that the `reservoir_size` could
easily be counted as one of those).

Nevermind that the `WienerHopfEsn` ultimately breaks this with its
additional `ridge_regression` parameter.